feat: update MLX plugin configuration and add new options for host and port
The MLX plugin configuration has been updated to include new options for the host and port of the HTTP server. The default values for host and port are "127.0.0.1" and "8080" respectively. Additionally, an optional path for the trained adapter weights has been added as "ADAPTER_FILE". These changes allow for more flexibility and customization when using the MLX plugin.